Added View Implementation For Games

+++ src/main/java/com/example/gamestore/ConsoleRunner.java
@@ -7,6 +7,7 @@ import com.example.gamestore.entities.userEntities.LoginUserDTO;
 import com.example.gamestore.entities.userEntities.RegisterUserDTO;
 import com.example.gamestore.entities.userEntities.User;
 import com.example.gamestore.exeptions.GameNotFoundException;
+import com.example.gamestore.exeptions.UserValidationException;
 import com.example.gamestore.exeptions.ValidationException;
 import com.example.gamestore.services.GameService;
 import com.example.gamestore.services.UserService;
@@ -14,7 +15,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Scanner;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 
 @Component
@@ -51,12 +56,16 @@ public class ConsoleRunner implements CommandLineRunner {
                 yield String.format("Successfully logged in %s", user.getFullName());
             }
             case "Logout" -> {
-                //TODO: logout user
-                //right now this only ends the program, but it should also logout the current user
                 isRunning = false;
+
+                validateLoggedUser();
+
+                this.userService.logout();
                 yield "Logged out";
             }
             case "AddGame" -> {
+                validateLoggedUser();
+                checkAdministratorRights();
                 AddGameDTO addGame = new AddGameDTO(commandData);
 
                 Game game = gameService.add(addGame);
@@ -64,20 +73,68 @@ public class ConsoleRunner implements CommandLineRunner {
                 yield String.format("Added %s", game.getTitle());
             }
             case "EditGame" -> {
-                //TODO: RETURN ONLY NEEDED ARRAY
-                //the whole array is send to the gameService, but the first two arguments are unnecessary
-                String title = this.gameService.edit(Integer.parseInt(commandData[1]), commandData);
+                validateLoggedUser();
+                checkAdministratorRights();
+
+                String[] editData = new String[commandData.length - 2];
+                System.arraycopy(commandData, 2, editData, 0, editData.length);
+
+                String title = this.gameService.edit(Integer.parseInt(commandData[1]), editData);
 
                 yield String.format("Edited %s", title);
-            } case "DeleteGame" -> {
+            }
+            case "DeleteGame" -> {
+                validateLoggedUser();
+                checkAdministratorRights();
+
                 DeleteGameDTO deleteGame = new DeleteGameDTO(commandData[1]);
 
                 String title = this.gameService.delete(deleteGame);
 
                 yield String.format("Deleted %s", title);
             }
+            //TODO: finish purchase logic
+            case "purchaseGame" -> {
+                validateLoggedUser();
+                Set<Game> games =  Arrays.stream(commandData).skip(1)
+                      .map(this.gameService::findByTitle)
+                      .collect(Collectors.toSet());
+                this.userService.purchaseGames(games);
+                yield "";
+            }
+            case "AllGames" -> {
+                List<Game> allGames = this.gameService.allGames();
+
+                StringBuilder sb = new StringBuilder();
+                allGames.forEach(g -> sb.append(String.format("%s %.2f%n", g.getTitle(), g.getPrice())));
+
+                yield sb.toString();
+            }
+            case "DetailGame" -> {
+                Game game = this.gameService.findByTitle(commandData[1]);
+
+                if (game == null) {
+                    throw new GameNotFoundException("No such game.");
+                }
+
+                yield String.format("Title: %s%nPrice: %.2f%nDescription: %s%nRelease date: %s",
+                        game.getTitle(), game.getPrice(), game.getDescription(), game.getReleaseDate());
+
+            }
             default -> "invalid command";
         };
+    }
+
+    private void checkAdministratorRights() throws UserValidationException {
+        if (!this.userService.getCurrentLoggedUser().isAdministrator()) {
+            throw new UserValidationException("You don't have rights to add/change games");
+        }
+    }
+
+    private void validateLoggedUser() throws UserValidationException {
+        if (this.userService.getCurrentLoggedUser() == null) {
+            throw new UserValidationException("There is no user currently logged in.");
+        }
     }
 
     @Override
@@ -89,7 +146,7 @@ public class ConsoleRunner implements CommandLineRunner {
 
             try {
                 System.out.println(execute(command));
-            } catch (ValidationException | GameNotFoundException e) {
+            } catch (ValidationException | GameNotFoundException | UserValidationException e) {
                 System.out.println(e.getMessage());
             }
         }
